python/bg_remove_service/requirements.txt
fastapi>=0.116,<1
uvicorn[standard]>=0.35,<1
python-multipart>=0.0.20,<1
pillow>=11,<12
httpx>=0.27,<1
anyio>=4.7,<5
# Match the torch line currently installed in ai-env; on a fresh machine, install the appropriate CUDA build first.
torch>=2.11,<2.12
torchvision>=0.26,<0.27
transformers>=4.55,<4.56
huggingface-hub>=0.34,<1
kornia>=0.8,<1
einops>=0.8,<1
timm>=1,<2
